+++ app/src/main/java/com/codesoom/assignment/controllers/SessionControllerErrorAdvice.java
@@ -1,0 +1,26 @@
+package com.codesoom.assignment.controllers;
+
+import com.codesoom.assignment.dto.ErrorResponse;
+import com.codesoom.assignment.errors.InvalidTokenException;
+import com.codesoom.assignment.errors.LoginDataNotMatchedException;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ControllerAdvice
+@ResponseBody
+public class SessionControllerErrorAdvice {
+    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+    @ExceptionHandler(LoginDataNotMatchedException.class)
+    public ErrorResponse handleLoginDataNotMatchedException(LoginDataNotMatchedException e) {
+        return new ErrorResponse(e.getMessage());
+    }
+
+    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+    @ExceptionHandler(InvalidTokenException.class)
+    public ErrorResponse handleLoginDataNotMatchedException(InvalidTokenException e) {
+        return new ErrorResponse(e.getMessage());
+    }
+}
